+++ twitchirc/twitchirc/__init__.py
@@ -18,7 +18,8 @@
 from .patterns import *
 from .moderation import ModerationContainer
 from .messages import (ChannelMessage, PartMessage, Message, JoinMessage, NoticeMessage, PingMessage, auto_message,
-                       process_twitch_flags, PongMessage, WhisperMessage, ReconnectMessage)
+                       process_twitch_flags, PongMessage, WhisperMessage, ReconnectMessage, UserstateMessage,
+                       UsernoticeMessage, GlobalNoticeMessage)
 from .middleware import Event, AbstractMiddleware
 from .connection import Connection
 from .command import Command